Add files via upload
Update and Delete Functions added to Transaction and Budget

Update Function doesn't work exactly as it is intended to. It updates by the previous value but not the new one

Delete function in budgetTrigger does the opposite and what it should do and removes the money from the account balance instead of restoring it. Does nothing when a transaction is added. I suspect all of the trigger statements are executing at once maybe causing some sort of conflict.